+++ azext_iot/tests/dps/__init__.py
@@ -20,7 +20,10 @@ from azext_iot.tests import CaptureOutputLiveScenarioTest
 from azext_iot.common.certops import create_self_signed_certificate
 from azext_iot.common.shared import AuthenticationTypeDataplane
 from azext_iot.tests.test_constants import ResourceTypes
+from knack.log import get_logger
 
+
+logger = get_logger(__name__)
 DATAPLANE_AUTH_TYPES = [
     AuthenticationTypeDataplane.key.value,
     AuthenticationTypeDataplane.login.value,
@@ -29,6 +32,10 @@ DATAPLANE_AUTH_TYPES = [
 
 CERT_NAME = "aziotcli"
 CERT_PATH = "aziotcli-cert.pem"
+KEY_PATH = "aziotcli-key.pem"
+SECONDARY_CERT_NAME = "aziotcli2"
+SECONDARY_CERT_PATH = "aziotcli2-cert.pem"
+SECONDARY_KEY_PATH = "aziotcli2-key.pem"
 WEBHOOK_URL = "https://www.test.test"
 API_VERSION = "2019-03-31"
 
@@ -37,6 +44,7 @@ PREFIX_EDGE_DEVICE = "test-edge-device-"
 PREFIX_INDIVIDUAL_ENROLLMENT = "test-enrollment-"
 PREFIX_GROUP_ENROLLMENT = "test-groupenroll-"
 USER_ROLE = "Device Provisioning Service Data Contributor"
+HUB_USER_ROLE = "IoT Hub Data Contributor"
 MAX_HUB_RETRIES = 3
 
 TEST_ENDORSEMENT_KEY = (
@@ -59,11 +67,12 @@ MAX_RBAC_ASSIGNMENT_TRIES = settings.env.azext_iot_rbac_max_tries if settings.en
 
 
 class IoTDPSLiveScenarioTest(CaptureOutputLiveScenarioTest):
-    def __init__(self, test_scenario):
+    def __init__(self, test_scenario, cert_only: bool = True):
         assert test_scenario
         self.entity_rg = ENTITY_RG
         self.entity_dps_name = ENTITY_DPS_NAME
         self.entity_hub_name = ENTITY_HUB_NAME
+        self.tracked_certs = []
         super(IoTDPSLiveScenarioTest, self).__init__(test_scenario)
 
         # Create resources if needed
@@ -87,15 +96,13 @@ class IoTDPSLiveScenarioTest(CaptureOutputLiveScenarioTest):
             rtype=ResourceTypes.hub.value,
             test_tag=test_scenario
         )
+        self.dps_cstring = self.get_dps_cstring()
+        self.hub_cstring = self.get_hub_cstring()
         self._ensure_dps_hub_link()
         self._cleanup_enrollments()
-        self.dps_cstring = self.get_dps_cstring()
 
         # Create the test certificate
-        output_dir = os.getcwd()
-        create_self_signed_certificate(
-            subject=CERT_NAME, valid_days=1, cert_output_dir=output_dir, cert_only=True
-        )
+        self.thumbprint = self.create_test_cert(cert_only=cert_only)
 
         # Kwargs
         base_enrollment_props = {
@@ -114,6 +121,21 @@ class IoTDPSLiveScenarioTest(CaptureOutputLiveScenarioTest):
 
         # Other variables for DPS testing
         self.hub_host_name = "{}.azure-devices.net".format(ENTITY_HUB_NAME)
+
+    def create_test_cert(self, subject=CERT_NAME, cert_only=True, file_prefix=None):
+        output_dir = os.getcwd()
+        thumbprint = create_self_signed_certificate(
+            subject=subject,
+            valid_days=1,
+            cert_output_dir=output_dir,
+            cert_only=cert_only,
+            file_prefix=file_prefix,
+            sha_version=256
+        )["thumbprint"]
+        self.tracked_certs.append(CERT_PATH)
+        if not cert_only:
+            self.tracked_certs.append(KEY_PATH)
+        return thumbprint
 
     def create_dps(self):
         """Create a device provisioning service for testing purposes."""
@@ -191,7 +213,7 @@ class IoTDPSLiveScenarioTest(CaptureOutputLiveScenarioTest):
         if "{}.azure-devices.net".format(self.entity_hub_name) not in hub_names:
             self.cmd(
                 "iot dps linked-hub create --dps-name {} -g {} --connection-string {} --location {}".format(
-                    self.entity_dps_name, self.entity_rg, self.get_hub_cstring(), self.get_hub_region()
+                    self.entity_dps_name, self.entity_rg, self.hub_cstring, self.get_hub_region()
                 )
             )
 
@@ -221,14 +243,33 @@ class IoTDPSLiveScenarioTest(CaptureOutputLiveScenarioTest):
                     )
                 )
 
-        self.cmd(
-            "iot dps enrollment list --dps-name {} -g  {}".format(self.entity_dps_name, self.entity_rg),
-            checks=self.is_empty(),
-        )
-        self.cmd(
-            "iot dps enrollment-group list --dps-name {} -g  {}".format(self.entity_dps_name, self.entity_rg),
-            checks=self.is_empty(),
-        )
+        # Check again only if needed
+        if enrollments:
+            self.cmd(
+                "iot dps enrollment list --dps-name {} -g  {}".format(self.entity_dps_name, self.entity_rg),
+                checks=self.is_empty(),
+            )
+
+        if enrollment_groups:
+            self.cmd(
+                "iot dps enrollment-group list --dps-name {} -g  {}".format(self.entity_dps_name, self.entity_rg),
+                checks=self.is_empty(),
+            )
+
+    def check_hub_device(self, device: str, auth_type: str, key: str = None, thumbprint: str = None):
+        """Helper method to check whether a device exists in a hub."""
+
+        device_auth = self.cmd(
+            "iot hub device-identity show -l {} -d {}".format(
+                self.hub_cstring,
+                device,
+            )
+        ).get_output_in_json()["authentication"]
+        assert auth_type == device_auth["type"]
+        if key:
+            assert key == device_auth["symmetricKey"]["primaryKey"]
+        if thumbprint:
+            assert thumbprint == device_auth["x509Thumbprint"]["primaryThumbprint"]
 
     def generate_device_names(self, count=1, edge=False):
         names = [
@@ -244,7 +285,7 @@ class IoTDPSLiveScenarioTest(CaptureOutputLiveScenarioTest):
             self.create_random_name(
                 prefix=PREFIX_INDIVIDUAL_ENROLLMENT if not group else PREFIX_GROUP_ENROLLMENT, length=48
             )
-            for i in range(count)
+            for _ in range(count)
         ]
         return names
 
@@ -266,6 +307,13 @@ class IoTDPSLiveScenarioTest(CaptureOutputLiveScenarioTest):
                 self.entity_dps_name, self.entity_rg, policy
             )
         ).get_output_in_json()["connectionString"]
+
+    def get_dps_id_scope(self):
+        return self.cmd(
+            "iot dps show -n {} -g {}".format(
+                self.entity_dps_name, self.entity_rg
+            )
+        ).get_output_in_json()["properties"]["idScope"]
 
     def set_cmd_auth_type(self, command: str, auth_type: str) -> str:
         if auth_type not in DATAPLANE_AUTH_TYPES:
@@ -289,8 +337,12 @@ class IoTDPSLiveScenarioTest(CaptureOutputLiveScenarioTest):
     @pytest.fixture(scope='class', autouse=True)
     def tearDownSuite(self):
         yield None
-        if os.path.exists(CERT_PATH):
-            os.remove(CERT_PATH)
+        for cert in self.tracked_certs:
+            if os.path.exists(cert):
+                try:
+                    os.remove(cert)
+                except OSError as e:
+                    logger.error(f"Failed to remove {cert}. {e}")
         if not settings.env.azext_iot_testhub:
             self.cmd(
                 "iot hub delete --name {} --resource-group {}".format(